Non-conflicting changes are retained. Conflicting paths keep the reconciled main state because it is the current integrated product line, while the source commit remains preserved as merge ancestry.

無衝突嘅改動照收；有衝突嘅路徑保留已整合嘅 main 狀態。來源提交仍然留喺合併歷史，唔會扮失蹤。

+++ llama/llama.cpp/src/llama-vocab.cpp
@@ -883,7 +883,7 @@ struct llm_tokenizer_ugm {
                     const double challenger_score = current_best.score_sum + token_score;
                     struct best_tokenization & current_champ = tokenization_results[prefix_offset];
                     if (challenger_score > current_champ.score_sum) {
-                        struct best_tokenization challenger = { token_id, input_offset, (float) challenger_score };
+                        struct best_tokenization challenger = { token_id, input_offset, challenger_score };
                         current_champ = challenger;
                     }
                 }
